demo_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
# Use the published SDK
x402-notify>=0.1.0
